Added AMC data import
Removed error in LANL naming convention which was causing a plotting error

Commented out downloading of static files. Causing read connection error on my end when date wasn't and files had to be downloaded.

Instead static files are just being read in. Ideally the static ones do not change so this should be fine for the future.